+++ lib/jira_add_on.dart
@@ -2,5 +2,6 @@ export './src/config.dart';
 export './src/domain/domain.dart';
 export './src/extensions/extensions.dart';
 export './src/models/models.dart';
+export './src/services/git.dart';
 export './src/services/jira.dart';
 export './src/utils/utils.dart';