+++ lib/src/models/project.dart
@@ -22,4 +22,9 @@ class Project {
       'name': name,
     };
   }
+
+  @override
+  String toString() {
+    return name;
+  }
 }