+++ lib/src/utils/utils.dart
@@ -1,1 +1,2 @@
 export './basic.dart';
+export './command.dart';